% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{createDurationSettings}
\alias{createDurationSettings}
\title{Create DurationSettings}
\usage{
createDurationSettings()
}
\value{
A \link{DurationSettings} R6 object.
}
\description{
Create DurationSettings
}
