% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{createPathwaySettings}
\alias{createPathwaySettings}
\title{Create PathwaySettings}
\usage{
createPathwaySettings(minFrequency = 1L, maxPathLength = 5L)
}
\arguments{
\item{minFrequency}{Integer. Minimum frequency to include a pathway in results (default 1).}

\item{maxPathLength}{Integer. Maximum number of treatment lines per pathway string (default 5).}
}
\value{
A \link{PathwaySettings} R6 object.
}
\description{
Create PathwaySettings
}
