% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{createSituationCountSettings}
\alias{createSituationCountSettings}
\title{Create SituationCountSettings}
\usage{
createSituationCountSettings()
}
\value{
A \link{SituationCountSettings} R6 object.
}
\description{
Create SituationCountSettings
}
