% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{createSituationTimeSettings}
\alias{createSituationTimeSettings}
\title{Create SituationTimeSettings}
\usage{
createSituationTimeSettings(maxLine = 5L)
}
\arguments{
\item{maxLine}{Integer. Maximum therapy line to include in "Time to Therapy Line"
output (default 5).}
}
\value{
A \link{SituationTimeSettings} R6 object.
}
\description{
Create SituationTimeSettings
}
