% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugHistory.R
\name{createTxHistoryTable}
\alias{createTxHistoryTable}
\title{Make the treatment history table as a temp in the database}
\usage{
createTxHistoryTable(
  connection,
  workDatabaseSchema,
  cohortTable,
  targetCohortId,
  eventCohortIds,
  gapDays = 32,
  tableName,
  queryMethod = "default"
)
}
\arguments{
\item{connection}{the connection to an omop database}

\item{workDatabaseSchema}{the database schema where results and cohorts can read and written}

\item{cohortTable}{the table where the cohort table is located}

\item{targetCohortId}{the id of the cohort in the cohort table}

\item{eventCohortIds}{the ids of the event cohort in the cohort table}

\item{gapDays}{the gap between eras to collapse on, defaults to 32}

\item{tableName}{the name of the table to save the treatment history table}

\item{queryMethod}{the methodology used to generate treatment history table. current options are 'default' (method used in ATLAS Pathways, without collapse rules for different tx cohorts) or 'atlas' (original ATLAS method)}
}
\description{
Make the treatment history table as a temp in the database
}
