% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxAnalysis.R
\name{DurationSettings}
\alias{DurationSettings}
\title{DurationSettings}
\description{
Placeholder settings class for treatment duration analysis. Currently
requires no configuration; instantiation enables the duration module.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DurationSettings-new}{\code{DurationSettings$new()}}
\item \href{#method-DurationSettings-clone}{\code{DurationSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DurationSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-DurationSettings-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DurationSettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationSettings$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DurationSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DurationSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
