% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howLong.R
\name{howLongAreDrugEras}
\alias{howLongAreDrugEras}
\title{How long is any drug era}
\usage{
howLongAreDrugEras(th)
}
\arguments{
\item{th}{the treatment history table}
}
\value{
a tibble summarizing the duration of eras using min, p25
p50, p75 and max values
}
\description{
How long is any drug era
}
