% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howLong.R
\name{howLongCombination}
\alias{howLongCombination}
\title{How long until start of combo treatment}
\usage{
howLongCombination(th)
}
\arguments{
\item{th}{the treatment history table}
}
\value{
a tibble summarizing the duration until start of combo treatment using min, p25
p50, p75 and max values
}
\description{
How long until start of combo treatment
}
