% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howLong.R
\name{howLongDiscontinuation}
\alias{howLongDiscontinuation}
\title{How long for patients until discontinuation}
\usage{
howLongDiscontinuation(th, days = 365)
}
\arguments{
\item{th}{the treatment history table}

\item{days}{the maximum number of days before stopping to consider an event a discontinuation}
}
\value{
a tibble summarizing the time to discontinuation using min, p25
p50, p75 and max values
}
\description{
How long for patients until discontinuation
}
