% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howLong.R
\name{howLongFirstTx}
\alias{howLongFirstTx}
\title{How long until first treatment}
\usage{
howLongFirstTx(th)
}
\arguments{
\item{th}{the treatment history table}
}
\value{
a tibble summarizing the duration until first treatment using min, p25
p50, p75 and max values
}
\description{
How long until first treatment
}
