% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howLong.R
\name{howLongOnlyTx}
\alias{howLongOnlyTx}
\title{How long only treatment}
\usage{
howLongOnlyTx(th)
}
\arguments{
\item{th}{the treatment history table}
}
\value{
a tibble summarizing the duration until next treatment using min, p25
p50, p75 and max values
}
\description{
How long only treatment
}
