% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howLong.R
\name{howLongTxLine}
\alias{howLongTxLine}
\title{How long is a specific treatment line}
\usage{
howLongTxLine(th, line = 1)
}
\arguments{
\item{th}{the treatment history table}

\item{line}{the treatment line to consider, defaults to 1}
}
\value{
a tibble summarizing the duration of eras using min, p25
p50, p75 and max values
}
\description{
How long is a specific treatment line
}
