% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howMany.R
\name{howManyComboTx}
\alias{howManyComboTx}
\title{How many patients have combination treatments}
\usage{
howManyComboTx(th)
}
\arguments{
\item{th}{the treatment history table}
}
\value{
a tibble summarizing the number of persons and percentage with a treatment combinations
}
\description{
How many patients have combination treatments
}
