% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howMany.R
\name{howManyInterruptTx}
\alias{howManyInterruptTx}
\title{How many patients interrupt treatment}
\usage{
howManyInterruptTx(th, breaks = c(30, 60, 9999))
}
\arguments{
\item{th}{the treatment history table}

\item{breaks}{interruption breaks}
}
\value{
a tibble summarizing the number of persons and percentage with a treatment sequence and event order by interruption
}
\description{
How many patients interrupt treatment
}
