% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howMany.R
\name{howManyMultiTx}
\alias{howManyMultiTx}
\title{How many patients have multiple treatments}
\usage{
howManyMultiTx(th)
}
\arguments{
\item{th}{the treatment history table}
}
\value{
a tibble summarizing the number of persons and percentage with a treatment sequence and event order
}
\description{
How many patients have multiple treatments
}
