% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howMany.R
\name{howManyTxLines}
\alias{howManyTxLines}
\title{How many lines of treatment do patients have}
\usage{
howManyTxLines(th)
}
\arguments{
\item{th}{the treatment history table}
}
\description{
How many lines of treatment do patients have
}
