% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howMany.R
\name{howManyWithEachTxInLine}
\alias{howManyWithEachTxInLine}
\title{How many had each treatment in a given line}
\usage{
howManyWithEachTxInLine(th, txLine)
}
\arguments{
\item{th}{the treatment history table}

\item{txLine}{the line of treatment (integer >= 1)}
}
\description{
How many had each treatment in a given line
}
