% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxAnalysis.R
\name{SituationCountSettings}
\alias{SituationCountSettings}
\title{SituationCountSettings}
\description{
Placeholder settings class for treatment situation count analysis. Currently
requires no configuration; instantiation enables the situation counts module.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SituationCountSettings-new}{\code{SituationCountSettings$new()}}
\item \href{#method-SituationCountSettings-clone}{\code{SituationCountSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SituationCountSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-SituationCountSettings-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SituationCountSettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SituationCountSettings$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SituationCountSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SituationCountSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SituationCountSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
