% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxAnalysis.R
\name{SituationTimeSettings}
\alias{SituationTimeSettings}
\title{SituationTimeSettings}
\description{
Settings for time-to-event analysis of treatment situations.
}
\details{
\code{maxLine} caps the maximum therapy line index included in the output.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{maxLine}}{Integer. Maximum therapy line index (read/write).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SituationTimeSettings-new}{\code{SituationTimeSettings$new()}}
\item \href{#method-SituationTimeSettings-clone}{\code{SituationTimeSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SituationTimeSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-SituationTimeSettings-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new SituationTimeSettings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SituationTimeSettings$new(maxLine = 5L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maxLine}}{Integer. Maximum therapy line to include in "Time to Therapy Line" output (default 5).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SituationTimeSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SituationTimeSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SituationTimeSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
