% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{truncateByEvent}
\alias{truncateByEvent}
\title{Truncate th to certain number of events}
\usage{
truncateByEvent(th, eventSeqLimit = 2)
}
\arguments{
\item{th}{the treatment history table}

\item{eventSeqLimit}{the cutoff of eligible event sequences in the treatment
history table. Defaults to 2}
}
\value{
the treatment history table filtered by the event seq limit
}
\description{
Truncate th to certain number of events
}
