% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{truncateByTime}
\alias{truncateByTime}
\title{Truncate th to certain number of events}
\usage{
truncateByTime(th, dayLimit = 365)
}
\arguments{
\item{th}{the treatment history table}

\item{dayLimit}{the cutoff of days considered in the treatment history table
Defaults to 365 (approx 1 year)}
}
\value{
the treatment history table filtered by the day limit
}
\description{
Truncate th to certain number of events
}
