% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howLong.R
\name{howLongDiscontinuation}
\alias{howLongDiscontinuation}
\title{How for patients until discontinuation}
\usage{
howLongDiscontinuation(th, threshold = 60, days = 365)
}
\arguments{
\item{th}{the treatment history table}
}
\value{
a tibble summarizing the time to discontinuation using min, p25
p50, p75 and max values
}
\description{
How for patients until discontinuation
}
