% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{viewSankey}
\alias{viewSankey}
\title{Make the sankey diagram for the treatment history}
\usage{
viewSankey(th, maxPathLength = 2)
}
\arguments{
\item{th}{the treatment history table}

\item{maxPathLength}{the maximum number of paths in the sequence}
}
\value{
a sankey diagram
}
\description{
Make the sankey diagram for the treatment history
}
